Move convertors of VmmActionErrors to VmmController
